requirements.txt
streamlit==1.21.0
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
seaborn==0.12.2
python-dotenv==1.0.0
neo4j==5.7.0
snowflake-connector-python==2.9.0
scikit-learn==1.4.2
imbalanced-learn==0.12.3